Added portal building to return to nether
This is still very glitchy. I can't test it a lot right now because I am too close to my base. Once the player goes through the portal, I will make it baritone efly to the highway coords, tower up to the highway, and backtrack until they are at the blue ice.